mevlut

+++ Business/Abstract/IRequestService.cs
@@ -10,6 +10,6 @@ namespace Business.Abstract
         List<Request> GetAll();
         //kategori id sine  göre listele
         List<Request> GetByCategoryId(int id);
-        List<Request> GetByCollectedAid(double collectedAid);
+        List<Request> GetByCollectedAid(int collectedAid);
     }
 }
